refactor: correccion componente usuarios
Se estaba repitiendo el componente de lista de usuarios en el componente de lista de areas, se ha corregido para que cada uno use su propio componente.
