+++ resources/js/components/areaInfoList.jsx
@@ -3,8 +3,7 @@ import { useLoadStatus } from "./context/loadContext";
 import { themeByProject } from "./utils/theme";
 import {  usePage } from "@inertiajs/react";
 
-
-function UserInfoList ({data}){
+function AreaInfoList ({data}){
     const { props } = usePage();
     const proyecto = props?.auth?.user?.proyecto || 'AZZU';
     const theme = themeByProject[proyecto];
@@ -14,19 +13,19 @@ function UserInfoList ({data}){
     return (
         <div className="grid auto-rows-min gap-4 md:grid-cols-3">
           <div className="border relative overflow-hidden rounded-xl border bg-white dark:bg-gray-800 p-4">
-            <h4 className="text-sm font-semibold text-gray-500 dark:text-gray-300">Total de usuarios</h4>
+            <h4 className="text-sm font-semibold text-gray-500 dark:text-gray-300">Total de areas</h4>
             <p className="mt-2 text-3xl font-bold text-gray-900 dark:text-white">{totalUsuarios}</p>
           </div>
 
           <div className="border relative overflow-hidden rounded-xl border bg-white dark:bg-gray-800 p-4">
-            <h4 className="text-sm font-semibold text-gray-500 dark:text-gray-300">Última usuario creado</h4>
+            <h4 className="text-sm font-semibold text-gray-500 dark:text-gray-300">Última areas creada</h4>
             <p className="mt-1 text-base font-medium text-gray-900 dark:text-white">
               {ultimoUsuario?.name || 'N/A'} <br/> {ultimoUsuario?.email}
             </p>
           </div>
 
           <div className="border relative overflow-hidden rounded-xl border bg-white dark:bg-gray-800 p-4">
-            <h4  className="text-sm font-semibold text-gray-500 dark:text-gray-300">usuarios activos</h4>
+            <h4  className="text-sm font-semibold text-gray-500 dark:text-gray-300">areas activas</h4>
             <p className={`mt-2 text-3xl font-bold ${theme.text}`}>
               {totalUsuarios} 
             </p>
@@ -37,4 +36,4 @@ function UserInfoList ({data}){
     );
 }
 
-export default UserInfoList
+export default AreaInfoList
